+++ src/WebApiBolierplate/Business.Lib/UserLib.cs
@@ -1,4 +1,5 @@
-﻿using Core.Constants;
+﻿using Business.Lib.Core;
+using Core.Constants;
 using Model;
 using System;