+++ src/WebApiBolierplate/Core.Constants/Config.cs
@@ -30,6 +30,7 @@
             public static string Key;
             public static string Issuer;
             public static string Audience;
+            public static int DaysValid;
         }
     }
 }