+++ src/WebApiBolierplate/Core.Constants/InfoTexts.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Core.Constants
+{
+    public static class InfoTexts
+    {
+        public const string Success = "Success";
+    }
+}